# docs/openapi-minecraft.yaml
openapi: 3.1.0
info:
  title: Minecraft MCP API
  description: REST API exposed by the Minecraft Fabric mod on port 7070. Wraps world manipulation, build task management, and prefab placement.
  version: 0.0.1
servers:
  - url: http://localhost:7070
    description: Local Minecraft server
...
paths:
  /api/area-locks:
    post:
      tags: [area-locks]
      summary: Reserve a cuboid
      operationId: acquireAreaLock
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [bounds]
              properties:
                bounds:
                  $ref: "#/components/schemas/AreaBounds"
                world:
                  $ref: "#/components/schemas/World"
                label:
                  type: string
      responses:
        "200":
          $ref: "#/components/responses/AreaLockGranted"
        "400":
          $ref: "#/components/responses/BadRequest"
        "409":
          $ref: "#/components/responses/AreaLockConflict"
    get:
      tags: [area-locks]
      summary: List reservations without disclosing tokens
      operationId: listAreaLocks
      description: Filter by world and optionally all six inclusive bounds coordinates. Does not renew reservations.
      parameters:
        - name: world
          in: query
          schema:
            type: string
        - name: min_x
          in: query
          schema:
            type: integer
            format: int32
        - name: min_y
          in: query
          schema:
            type: integer
            format: int32
        - name: min_z
          in: query
          schema:
            type: integer
            format: int32
        - name: max_x
          in: query
          schema:
            type: integer
            format: int32
        - name: max_y
          in: query
          schema:
            type: integer
            format: int32
        - name: max_z
          in: query
          schema:
            type: integer
            format: int32
      responses:
        "200":
          description: Active reservations
          content:
            application/json:
              schema:
                type: object
                properties:
                  locks:
                    type: array
                    items:
                      $ref: "#/components/schemas/AreaLockMetadata"
        "400":
          $ref: "#/components/responses/BadRequest"
  /api/area-locks/{lock_id}:
    parameters:
      - $ref: "#/components/parameters/AreaLockPathId"
    patch:
      tags: [area-locks]
      summary: Renew or resize a reservation
      operationId: updateAreaLock
      description: An empty object renews. Replacement bounds resize atomically; a failed resize preserves the old bounds and expiry. The world cannot change.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                bounds:
                  $ref: "#/components/schemas/AreaBounds"
      responses:
        "200":
          $ref: "#/components/responses/AreaLockGranted"
        "400":
          $ref: "#/components/responses/BadRequest"
        "409":
          $ref: "#/components/responses/AreaLockConflict"
    delete:
      tags: [area-locks]
      summary: Release a reservation (idempotent)
      operationId: releaseAreaLock
      responses:
        "200":
          description: Reservation released or already absent
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean


  # ── System ──────────────────────────────────────────────────────────────────
...
  /api/world/blocks/set:
    parameters:
      - $ref: "#/components/parameters/AreaLockHeader"
    post:
      tags: [blocks]
      summary: Set a 3-D array of blocks
      operationId: setBlocks
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BlockSetRequest"
      responses:
        "409":
          $ref: "#/components/responses/AreaLockConflict"
        "200":
          description: Blocks placed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BlockSetResponse"
        "500":
          $ref: "#/components/responses/ServerError"
  /api/world/blocks/chunk:
    post:
      tags: [blocks]
      summary: Read a 3-D chunk of blocks
      operationId: getBlocksChunk
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ChunkRequest"
      responses:
        "200":
          description: Block chunk data
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ChunkResponse"
        "500":
          $ref: "#/components/responses/ServerError"
  /api/world/blocks/fill:
    parameters:
      - $ref: "#/components/parameters/AreaLockHeader"
    post:
      tags: [blocks]
      summary: Fill a rectangular region with a single block type
      operationId: fillBox
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/FillBoxRequest"
      responses:
        "409":
          $ref: "#/components/responses/AreaLockConflict"
        "200":
          description: Fill completed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FillResponse"
        "500":
          $ref: "#/components/responses/ServerError"
  /api/world/blocks/heightmap:
    post:
      tags: [blocks]
      summary: Get heightmap for a region
      operationId: getHeightmap
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/HeightmapRequest"
      responses:
        "200":
          description: Heightmap data
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HeightmapResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "500":
          $ref: "#/components/responses/ServerError"
  /api/world/blocks/heightmap/preview:
    post:
      tags: [blocks]
      summary: Render isometric heightmap preview as PNG
      operationId: previewHeightmap
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/HeightmapPreviewRequest"
      responses:
        "200":
          description: PNG image
          content:
            image/png:
              schema:
                type: string
                format: binary
        "204":
          description: No blocks in region
        "400":
          $ref: "#/components/responses/BadRequest"
        "500":
          $ref: "#/components/responses/ServerError"

  # ── Messages ─────────────────────────────────────────────────────────────────
...
components:
  parameters:
    AreaLockHeader:
      name: X-Area-Lock-Id
      in: header
      required: false
      description: Optional reservation token. Writes must fit inside its world and cuboid. Unknown or expired tokens fail; absence only permits writes outside reserved areas. Execute/replay carries the token into each task and stops on a lock failure.
      schema:
        type: string
    AreaLockPathId:
      name: lock_id
      in: path
      required: true
      schema:
        type: string
...
  responses:
    AreaLockGranted:
      description: Reservation including its usable token and idle expiry
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/AreaLock"
    AreaLockConflict:
      description: Conflicting reservation, write outside its reservation, or unknown/expired token. No blocks are written by the rejected operation. Queued task failures are reported in build status instead.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/AreaLockError"
    BadRequest:
      description: Invalid request
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
...
    ServerError:
      description: Internal server error
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
  schemas:
    AreaBounds:
      type: object
      description: Inclusive block bounds; each minimum must not exceed its maximum.
      required: [min_x, min_y, min_z, max_x, max_y, max_z]
      properties:
        min_x:
          type: integer
          format: int32
        min_y:
          type: integer
          format: int32
        min_z:
          type: integer
          format: int32
        max_x:
          type: integer
          format: int32
        max_y:
          type: integer
          format: int32
        max_z:
          type: integer
          format: int32
    AreaLockMetadata:
      type: object
      required: [world, bounds, label, expires_at, idle_timeout_seconds]
      properties:
        world:
          type: string
        bounds:
          $ref: "#/components/schemas/AreaBounds"
        label:
          type: string
        expires_at:
          type: string
          format: date-time
        idle_timeout_seconds:
          type: integer
          default: 900
    AreaLock:
      allOf:
        - $ref: "#/components/schemas/AreaLockMetadata"
        - type: object
          required: [lock_id]
          properties:
            lock_id:
              type: string
              format: uuid
    AreaLockError:
      type: object
      required: [success, code, error]
      properties:
        success:
          type: boolean
          const: false
        code:
          type: string
          enum: [area_locked, outside_area_lock, invalid_area_lock]
        error:
          type: string
        reservation:
          $ref: "#/components/schemas/AreaLockMetadata"
        requested_world:
          type: string
        requested_bounds:
          type: object
          description: Inclusive operation bounds; empty for a no-op with a mismatched world.

    # ── Shared primitives ──────────────────────────────────────────────────────
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
    World:
      type: string
      default: minecraft:overworld
      example: minecraft:overworld
    Position3D:
      type: object
      properties:
        x:
          type: number
        y:
          type: number
        z:
          type: number
      required: [x, y, z]
    BlockData:
      type: object
      properties:
        block_name:
          type: string
          example: minecraft:stone
        block_states:
          type: object
          additionalProperties:
            type: string
          example:
            facing: north
      required: [block_name]
...
    HeightmapType:
      type: string
      enum: [WORLD_SURFACE, MOTION_BLOCKING, MOTION_BLOCKING_NO_LEAVES, OCEAN_FLOOR]
      default: WORLD_SURFACE
...
    ViewDirection:
      type: string
      enum: [south, west, north, east]
      default: south
...
    BlockSetRequest:
      type: object
      properties:
        world:
          $ref: "#/components/schemas/World"
        start_x:
          type: integer
        start_y:
          type: integer
        start_z:
          type: integer
        blocks:
          type: array
          description: 3-D array [x][y][z]; null entries are skipped
          items:
            type: array
            items:
              type: array
              items:
                oneOf:
                  - $ref: "#/components/schemas/BlockData"
                  - type: "null"
      required: [start_x, start_y, start_z, blocks]
    BlockSetResponse:
      type: object
      properties:
        success:
          type: boolean
        blocks_set:
          type: integer
        blocks_skipped:
          type: integer
        world:
          type: string
    ChunkRequest:
      type: object
      properties:
        world:
          $ref: "#/components/schemas/World"
        start_x:
          type: integer
        start_y:
          type: integer
        start_z:
          type: integer
        size_x:
          type: integer
        size_y:
          type: integer
        size_z:
          type: integer
      required: [start_x, start_y, start_z, size_x, size_y, size_z]
    ChunkResponse:
      type: object
      properties:
        success:
          type: boolean
        world:
          type: string
        start_position:
          $ref: "#/components/schemas/Position3D"
        size:
          $ref: "#/components/schemas/Position3D"
        blocks:
          type: array
          description: 3-D array [x][y][z]
          items:
            type: array
            items:
              type: array
              items:
                $ref: "#/components/schemas/BlockData"
    FillBoxRequest:
      type: object
      properties:
        world:
          $ref: "#/components/schemas/World"
        x1:
          type: integer
        y1:
          type: integer
        z1:
          type: integer
        x2:
          type: integer
        y2:
          type: integer
        z2:
          type: integer
        block_type:
          type: string
          example: minecraft:stone
        notify_neighbors:
          type: boolean
          default: false
      required: [x1, y1, z1, x2, y2, z2, block_type]
    FillResponse:
      type: object
      properties:
        success:
          type: boolean
        blocks_set:
          type: integer
        blocks_failed:
          type: integer
        total_blocks:
          type: integer
        world:
          type: string
        box_bounds:
          type: object
    HeightmapRequest:
      type: object
      properties:
        world:
          $ref: "#/components/schemas/World"
        x1:
          type: integer
        z1:
          type: integer
        x2:
          type: integer
        z2:
          type: integer
        heightmap_type:
          $ref: "#/components/schemas/HeightmapType"
      required: [x1, z1, x2, z2]
    HeightmapResponse:
      type: object
      properties:
        success:
          type: boolean
        world:
          type: string
        area_bounds:
          type: object
        size:
          type: object
          properties:
            width:
              type: integer
            height:
              type: integer
        heightmap_type:
          type: string
        height_range:
          type: object
          properties:
            min:
              type: integer
            max:
              type: integer
        heights:
          type: array
          description: 2-D array of height values
          items:
            type: array
            items:
              type: integer
    HeightmapPreviewRequest:
      allOf:
        - $ref: "#/components/schemas/HeightmapRequest"
        - type: object
          properties:
            iso_scale:
              type: integer
              minimum: 1
              maximum: 32
              default: 6
            view_direction:
              $ref: "#/components/schemas/ViewDirection"

    # ── Messages ──────────────────────────────────────────────────────────────
...